+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/setup/state/ClientMutableState.kt
@@ -18,28 +18,26 @@ package io.getstream.chat.android.client.setup.state
 
 import io.getstream.chat.android.client.models.ConnectionState
 import io.getstream.chat.android.client.models.InitializationState
-import io.getstream.chat.android.core.internal.InternalStreamChatApi
 
 /**
  * Mutable version of [ClientState]. The class makes possible to change state of the SDK. Should only be used
  * internally by the SDK.
  */
-@InternalStreamChatApi
-public interface ClientMutableState : ClientState {
+internal interface ClientMutableState : ClientState {
 
     /**
      * Sets the [ConnectionState]
      *
      * @param connectionState [ConnectionState]
      */
-    public fun setConnectionState(connectionState: ConnectionState)
+    fun setConnectionState(connectionState: ConnectionState)
 
     /**
      * Sets initialized
      *
      * @param state [InitializationState]
      */
-    public fun setInitializationState(state: InitializationState)
+    fun setInitializationState(state: InitializationState)
 
     /**
      * Clears the state of [ClientMutableState].